Update deployment process so that packages can be released independently
Untested, but the idea is that versions are still in the Makefile, but now separated per package. Releases are now triggered via tags, e.g. xRetry_v1.9.0 or xRetry.Reqnroll_v1.0.0-alpha1. The tag name prefix is what will trigger the release of that specific package.

Note: the version in the tag is for info only (git history & GH releases). The actual package version deployed will be what is in the makefile, and if it's a Reqnroll or SpecFlow plugin package, the xRetry version they depend on will also be from the Makefile.