Fix NU5128 dotnet pack warnings
It detects that lib\* assemblies are listed for multiple target frameworks, but dependencies are not separated by framework & warns about this. The current guidance is to separate dependencies by target framework, even if the dependencies are identical.

Note that I haven't applied the same change to xRetry.Specflow, since it didn't warn for that. That seems to be because the different files are under build\*, not lib\* or ref\*...

From the docs (https://learn.microsoft.com/en-us/nuget/reference/errors-and-warnings/nu5128):
If you are using a custom nuspec file, we recommend each TFM for which lib/ or ref/ assemblies exist should have a matching dependency group, even if the dependencies are the same as the next compatible TFM. For example, if a package contains netstandard1.0 and netstandard2.0 assemblies, and the dependencies are the same for both, we recommend both TFMs be listed as dependency groups with duplicate dependency items.

Note that the TFM identifier used in the assembly paths use a different format to the TFM identifier used in dependency groups. The warning message specifies the correct name to use in the dependency group. If your package does not have any dependencies for that target framework, use an empty group