Remove custom skipping on exceptions for xRetry.v3
xUnit.v3 now supports this natively, so we don't need to add it. Some tests are left in, but now using the xUnit.v3 way. This lets us check we aren't retrying skipped tests.

Docs are updated. The section is not entirely removed, but replaced with a message explaining how to upgrade from the old xRetry way to the new xUnit.v3 way of skipping at runtime.